Al-Ansaar Project/requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.46
WTForms==3.0.1
Flask-WTF==1.1.1
Flask-Login==0.6.0
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
python-dotenv==1.0.0
secure==0.3.0
requests==2.31.0
pytz==2021.3
